requirements.txt
pandas
scrapling[fetchers]
geocoder
pytz
timezonefinder
astropy
streamlit>=1.40.0
streamlit-js-eval
streamlit-searchbox
astroquery
PyGithub
PyYAML
reportlab